add support for doxygen file generation
