+++ ablateLibrary/finiteVolume/processes/twoPhaseEulerAdvection.hpp
@@ -9,6 +9,18 @@ namespace ablate::finiteVolume::processes {
 
 class TwoPhaseEulerAdvection : public Process {
    private:
+    struct DecodeDataStruct {
+        std::shared_ptr<ablate::eos::EOS> eosGas;
+        std::shared_ptr<ablate::eos::EOS> eosLiquid;
+        PetscReal ke;
+        PetscReal e;
+        PetscReal rho;
+        PetscReal Yg;
+        PetscReal Yl;
+        PetscInt dim;
+        PetscReal* vel;
+    };
+
     const std::shared_ptr<eos::EOS> eosGas;
     const std::shared_ptr<eos::EOS> eosLiquid;
     const std::shared_ptr<fluxCalculator::FluxCalculator> fluxCalculatorGasGas;
@@ -20,6 +32,8 @@ class TwoPhaseEulerAdvection : public Process {
         PetscReal g[3];  // default = {0.0, 0.0, 0.0};
     };
     Parameters parameters;
+
+    static PetscErrorCode FormFunction(SNES snes, Vec x, Vec F, void* ctx);
 
    public:
     static PetscErrorCode UpdateAuxTemperatureField2Gas(PetscReal time, PetscInt dim, const PetscFVCellGeom* cellGeom, const PetscInt uOff[], const PetscScalar* conservedValues, PetscScalar* auxField,
