+++ docs/_layouts/redirect.html
@@ -1,0 +1,5 @@
+<!DOCTYPE html>
+<meta charset="utf-8">
+<title>Redirecting to {{page.redirect_url}}</title>
+<meta http-equiv="refresh" content="0; URL={{page.redirect_url}}">
+<link rel="canonical" href="{{page.redirect_url}}">
