+++ docs/content/SourceDocumentation.md
@@ -1,0 +1,7 @@
+---
+title: Source Documentation 🔗
+layout: redirect
+redirect_url: /sourceDocs/
+has_children: false
+nav_order: 10
+---
